refactor: participant 패키지명, Participant 클래스명 변경 및 keepCard 메서드 추상메서드로 변경
