+++ src/main/java/domain/participant/Players.java
@@ -1,4 +1,4 @@
-package domain.participants;
+package domain.participant;
 
 import java.util.Collections;
 import java.util.List;
