--- src/main/java/domain/participants/Player.java
@@ -1,8 +1,0 @@
-package domain.participants;
-
-public class Player extends Participants {
-
-    public Player(String name, Hand hand) {
-        super(name, hand);
-    }
-}
