+++ src/test/java/BlackJackServiceTest.java
@@ -3,10 +3,10 @@ import domain.card.Pattern;
 import domain.card.Rank;
 import domain.game.Result;
 import domain.game.ResultInfo;
-import domain.participants.Dealer;
-import domain.participants.Hand;
-import domain.participants.Player;
-import domain.participants.Players;
+import domain.participant.Dealer;
+import domain.participant.Hand;
+import domain.participant.Player;
+import domain.participant.Players;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import service.BlackJackService;
